Update README.md
The code and dataset used to analyze the energy efficiency of buildings are contained in this repository. To predict heating and cooling loads depending on different building attributes, the project entails running SQL queries in R, manipulating and transforming data, and producing visualizations. Features like surface, wall, roof area, glass area, glass area distribution, and orientation of simulated building designs are all included in the dataset. The goal is to estimate the building's heating and cooling load using this data. By increasing building energy efficiency, the analysis hopes to lessen carbon emissions and offer insights into patterns of energy consumption.